# openapi.yaml
openapi: 3.0.3
info:
  title: ACS Multi-Vendor TR-069/CWMP REST API
  version: 1.0.0
  description: |
    REST API internal (Echo v5) untuk Auto Configuration Server (ACS) TR-069/CWMP
    multi-vendor (ZTE, Huawei, FiberHome, Nokia, C-Data, dst). Dikonsumsi oleh
    sistem BSS/OSS, portal NOC, dan integrasi internal lain — **BUKAN** endpoint
    yang dipakai CPE (ONT/ONU/router) itu sendiri.

    **PENTING — beda dengan endpoint CWMP:** perangkat CPE berkomunikasi dengan
    ACS lewat endpoint `/cwmp` (SOAP/XML, protokol TR-069/CWMP standar Broadband
    Forum, sesi stateful) yang **TIDAK** didokumentasikan di file ini sama
    sekali — beda protokol, beda tujuan, beda konsumen. Dokumen ini HANYA
    mencakup `/api/v1/*` (REST + JSON, kecuali `GET /api/v1/metrics` yang
    formatnya teks Prometheus).

    ## Autentikasi

    Hampir seluruh endpoint mewajibkan header `Authorization: Bearer <token>`,
    berupa salah satu dari:
    - **JWT sesi user** — didapat dari `POST /auth/login`, berlaku sesuai masa
      berlaku yang dikonfigurasi (`ACS_JWT_EXPIRY`, default 8 jam di `.env.example`).
    - **API token** — string berawalan `acs_...`, diterbitkan lewat
      `POST /auth/tokens` (ADMIN/SUPERADMIN), tidak pernah kedaluwarsa kecuali
      `expires_at` diisi.

    Empat endpoint publik tanpa autentikasi, semuanya READ-ONLY atau titik masuk
    login (tidak ada endpoint MUTASI tanpa autentikasi): `POST /auth/login`
...
    aman).

...

    ## RBAC

...

...

...
    skema `Task`.

    ## Server

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/tokens:
    post:
      operationId: issueAPIToken
      tags: [Auth]
      summary: Menerbitkan API token jangka panjang (`acs_...`)
      description: |
        ADMIN/SUPERADMIN. Non-superadmin hanya boleh menerbitkan token untuk
        tenant miliknya sendiri (`tenant_id` di body diabaikan/ditolak bila
        beda dari tenant actor). Nilai token PLAINTEXT hanya ditampilkan SEKALI
        di response ini — hanya hash-nya yang disimpan, tidak bisa diambil ulang.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAPITokenRequest'
      responses:
        '201':
          description: Token diterbitkan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueAPITokenResponse'
        '400':
          description: Payload tidak valid atau `name` kosong.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role tidak cukup, atau non-superadmin mencoba menerbitkan token untuk tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    get:
      operationId: listAPITokens
      tags: [Auth]
      summary: Daftar API token yang pernah diterbitkan (untuk dikelola/dicabut)
      description: |
        ADMIN/SUPERADMIN — gating role SAMA seperti penerbitan token
        (`POST /auth/tokens`), bukan self-service ke semua role. Non-superadmin
        SELALU dipaksa ke tenant miliknya sendiri (parameter `tenant_id` diabaikan
        bila diisi tenant lain) dan melihat SELURUH token milik tenant tsb —
        bukan cuma token yang diterbitkan actor sendiri — supaya ADMIN bisa
        mencabut token yang diterbitkan ADMIN lain di tenant yang sama (mis.
        karyawan penerbit token sudah resign). Nilai `token` plaintext TIDAK
        pernah muncul di sini (hanya saat `POST /auth/tokens` pertama kali).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/tenantIdQuery'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Daftar API token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APITokenListResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role tidak cukup.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/tokens/{id}:
    delete:
      operationId: revokeAPIToken
      tags: [Auth]
      summary: Mencabut (revoke) satu API token
      description: |
        ADMIN/SUPERADMIN. Non-superadmin hanya boleh mencabut token milik
        tenant-nya sendiri — mencoba mencabut token tenant lain ditolak `403`,
        TIDAK PERNAH diam-diam berhasil. **Catatan jujur soal enumerasi**: `id`
        tidak ada sama sekali → `404`, `id` ada tapi milik tenant lain → `403`
        — kedua kode ini berbeda, sehingga non-superadmin YANG MENEBAK id
        secara berurutan bisa membedakan "id belum pernah dipakai" dari "id
        ini token milik tenant lain yang (pernah) ada". Diterima apa adanya
        (severity rendah — `id` adalah PK internal BIGINT sekuensial pada
        endpoint admin-only, bukan identifier publik, dan pola 404-vs-403 yang
        sama juga berlaku di endpoint lain di API ini), bukan dianggap
        kebocoran isi/kepemilikan token. Mencabut token yang sudah revoked
        sebelumnya tetap `204` (idempotent), bukan error.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Token berhasil dicabut (atau memang sudah revoked sebelumnya).
        '400':
          description: "`id` tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role tidak cukup, atau non-superadmin mencoba mencabut token tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Token tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/password:
    patch:
      operationId: changeOwnPassword
      tags: [Auth]
      summary: Ganti password akun sendiri
      description: |
        Semua role yang sudah login boleh memanggil ini (tidak ada gating role)
        — actor diambil dari JWT/API token, BUKAN dari `:id` path (beda dari
        `PATCH /users/:id/password` yang itu admin mereset password user
        LAIN). Wajib menyertakan `current_password` yang benar; tanpa verifikasi
        ini siapa pun yang membajak bearer token bisa mengunci pemilik akun asli
        secara permanen.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeOwnPasswordRequest'
      responses:
        '204':
          description: Password berhasil diganti.
        '400':
          description: Payload tidak valid, field kosong, atau `new_password` kurang dari 8 karakter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi, atau `current_password` salah.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================== REFS ==============================
...
  /users/{id}/password:
    patch:
      operationId: resetUserPassword
      tags: [Users]
      summary: Admin mereset password user lain
      description: |
        ADMIN/SUPERADMIN, scope sama seperti `PATCH /users/:id`. BEDA dari
        `PATCH /auth/password` (ganti password sendiri dgn verifikasi password
        lama) — di sini TIDAK PERLU tahu password lama user target. Password
        baru tidak pernah dicatat ke activity log. Sekalian mereset lockout
        brute-force (failed attempts + locked_until) milik user target.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetUserPasswordRequest' }
      responses:
        '204':
          description: Password berhasil direset.
        '400':
          description: "`id` tidak valid, payload tidak valid, `new_password` kosong atau kurang dari 8 karakter."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau ADMIN mencoba mereset password user tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /auth/oidc/login:
    get:
      operationId: oidcLogin
      tags: [Auth]
      summary: Mulai alur login OIDC (redirect ke Identity Provider)
      description: |
        Hanya aktif bila OIDC dikonfigurasi (`ACS_OIDC_*`). Endpoint browser
        redirect — BUKAN JSON. Mengembalikan `302` (`http.StatusFound`) ke
        authorization endpoint IdP. Tanpa autentikasi (titik masuk login).

        **Catatan implementasi (bukan spec):** parameter `state` saat ini masih
        nilai statis di handler (`oidcLogin`, ada `TODO` untuk generator state
        acak) — jangan diandalkan sebagai proteksi CSRF sampai TODO itu selesai.
      security: []
      responses:
        '302':
          description: Redirect ke Identity Provider.
        '500':
          description: OIDC tidak dikonfigurasi/gagal membangun authorization URL di instance ini (handler mengembalikan `500`, bukan `404`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/oidc/callback:
    get:
      operationId: oidcCallback
      tags: [Auth]
      summary: Callback OIDC — tukar authorization code jadi sesi
      description: |
        Dipanggil oleh browser setelah IdP meng-authorize user. Menukar
        authorization `code` menjadi ID token, lalu menautkan/menerbitkan user
        lokal. Tanpa autentikasi Bearer (identitas dibawa oleh `code`).

        Status implementasi saat ini (POC, lihat komentar di `oidcCallback`):
        handler mengembalikan **`200` JSON dengan bentuk yang SAMA PERSIS
        seperti `POST /auth/login`** (`access_token` + `user`), BUKAN redirect
        `302` — frontend menyimpan token lalu berpindah halaman sendiri.
        Handler hanya membaca query `code`; parameter `state` TIDAK dibaca
        sama sekali (belum divalidasi — lihat catatan CSRF di
        `GET /auth/oidc/login`), jadi sengaja tidak didokumentasikan sebagai
        parameter kontrak.
      security: []
      parameters:
        - { name: code, in: query, required: true, description: "Authorization code dari IdP.", schema: { type: string } }
      responses:
        '200':
          description: Login OIDC berhasil — bentuk response identik `POST /auth/login`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400':
          description: "Query `code` kosong/tidak ada."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: "`code` invalid, ID token gagal diverifikasi, atau user tidak bisa ditautkan."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================== REALTIME ==============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT sesi user (dari `POST /auth/login`) ATAU API token statis
        berawalan `acs_...` (dari `POST /auth/tokens`) — keduanya dikirim
        dengan cara yang sama lewat header `Authorization: Bearer <token>`.
  parameters:
    page:
      name: page
      in: query
      description: Halaman (1-based). Default 1.
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: page_size
      in: query
      description: Ukuran halaman. Default 50, maksimum 200 (nilai di luar rentang ini jatuh ke default 50).
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    tenantIdQuery:
      name: tenant_id
      in: query
      description: |
        Filter tenant. HANYA efektif untuk actor `SUPERADMIN` — actor role lain
        selalu dipaksa ke tenant miliknya sendiri, nilai ini diabaikan.
      schema: { type: integer, format: int64 }
...
    userIdPath:
      name: id
      in: path
      required: true
      description: ID numerik internal user.
      schema: { type: integer, format: int64 }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Bentuk error SERAGAM untuk seluruh endpoint (`echo.NewHTTPError`) —
        selalu `{"message": "..."}`, tidak ada kode error mesin terpisah dari
        `message` teks bebas saat ini (perhatikan: teks pesan sebagian besar
        Bahasa Indonesia, konsisten dengan gaya pesan aplikasi ini).
      properties:
        message:
          type: string
          example: "domain: data tidak ditemukan"
      required: [message]
...
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
        user:
          type: object
          properties:
            id: { type: integer, format: int64 }
            uuid: { type: string, format: uuid }
            username: { type: string }
            roles: { type: array, items: { type: string } }
            tenant_id: { type: integer, format: int64, nullable: true }
      required: [access_token, token_type, user]
    ChangeOwnPasswordRequest:
      type: object
      properties:
        current_password: { type: string, format: password }
        new_password: { type: string, format: password, minLength: 8 }
      required: [current_password, new_password]
    IssueAPITokenRequest:
      type: object
      properties:
        name: { type: string, maxLength: 128, description: Nama/label token untuk memudahkan identifikasi di masa depan. }
        tenant_id:
          type: integer
          format: int64
          nullable: true
          description: "Non-superadmin: harus sama dengan tenant milik actor sendiri (atau dikosongkan agar default ke tenant actor)."
        expires_at: { type: string, format: date-time, nullable: true, description: "Kosongkan agar token tidak pernah kedaluwarsa." }
      required: [name]
    IssueAPITokenResponse:
      type: object
      description: "Nilai `token` PLAINTEXT hanya muncul di sini, SEKALI — tidak bisa diambil ulang lewat endpoint manapun setelahnya."
      properties:
        token: { type: string, example: "acs_XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX" }
        id: { type: integer, format: int64 }
        uuid: { type: string, format: uuid }
      required: [token, id, uuid]
    APIToken:
      type: object
      description: |
        Metadata satu API token (dari `GET /auth/tokens`). Nilai token
        PLAINTEXT tidak pernah ada di sini — hanya `token_hash` yang tersimpan
        di database, dan field itu sendiri disembunyikan dari response
        (`json:"-"` di Go, lihat domain/tenancy.go).
      properties:
        id: { type: integer, format: int64 }
        token_uuid: { type: string, format: uuid }
        user_id: { type: integer, format: int64, nullable: true, description: "User yang menerbitkan token ini." }
        tenant_id: { type: integer, format: int64, nullable: true }
        name: { type: string }
        scopes:
          type: object
          nullable: true
          description: "Belum ada jalur yang mengisi field ini (selalu `null` saat ini) — dicadangkan utk scope granular per-token di masa depan."
        expires_at: { type: string, format: date-time, nullable: true }
        revoked_at: { type: string, format: date-time, nullable: true, description: "NULL berarti masih aktif." }
        created_at: { type: string, format: date-time }
        created_by: { type: integer, format: int64, nullable: true }
        updated_at: { type: string, format: date-time }
        updated_by: { type: integer, format: int64, nullable: true }
      required: [id, token_uuid, name, created_at, updated_at]
    APITokenListResponse:
      type: object
      properties:
        data:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/APIToken' }
        total: { type: integer }
      required: [data, total]

    # ---------- Tenants ----------
...
    ResetUserPasswordRequest:
      type: object
      properties:
        new_password: { type: string, format: password, minLength: 8 }
      required: [new_password]
...